For the largest part, I don't know what I changed (including, of course, not knowing which of those unknown changes I changed back again).. Only ExerciseAO.vue seems to be affected, though. And it still seems to work alright.
I managed to get the video in AO and the picture in MI&Finished not to grow too much as the window grows, though.
And I added a click event to the back button in AO - and unsuccessfully tried to add one to the next button there as well (goes back instead of forward).